Algebra.Heyting.Free.Expr

Documentation

data Expr a

Heyting algebra expression.

Note: this type doesn't have Heyting instance, as its Eq and Ord are structural.

Constructors

Var a

Bottom

Top

(Expr a) :/\: (Expr a) infixr 6

(Expr a) :\/: (Expr a) infixr 5

(Expr a) :=>: (Expr a) infixr 4

proofSearch :: forall a. Ord a => Expr a -> Bool

Decide whether x :: Expr a is provable.

Note: this doesn't construct a proof term, but merely returns a Bool.