Algebra.Lattice

Description

In mathematics, a lattice is a partially ordered set in which every two elements have a unique supremum (also called a least upper bound or join) and a unique infimum (also called a greatest lower bound or meet).

In this module lattices are defined using meet and join operators, as it's constructive one.

Unbounded lattices

class Lattice a where

An algebraic structure with joins and meets.

See http://en.wikipedia.org/wiki/Lattice_(order) and http://en.wikipedia.org/wiki/Absorption_law.

Lattice is very symmetric, which is seen from the laws:

Associativity

x \/ (y \/ z) ≡ (x \/ y) \/ z
x /\ (y /\ z) ≡ (x /\ y) /\ z

Commutativity

x \/ y ≡ y \/ x
x /\ y ≡ y /\ x

Idempotency

x \/ x ≡ x
x /\ x ≡ x

Absorption

a \/ (a /\ b) ≡ a
a /\ (a \/ b) ≡ a

Methods

(\/) :: a -> a -> a infixr 5

join

(/\) :: a -> a -> a infixr 6

meet

joinLeq :: (Eq a, Lattice a) => a -> a -> Bool

The partial ordering induced by the join-semilattice structure

joins1 :: (Lattice a, Foldable1 f) => f a -> a

The join of at a list of join-semilattice elements (of length at least one)

meetLeq :: (Eq a, Lattice a) => a -> a -> Bool

meets1 :: (Lattice a, Foldable1 f) => f a -> a

The meet of at a list of meet-semilattice elements (of length at least one)

Bounded lattices

class Lattice a => BoundedJoinSemiLattice a where

A join-semilattice with an identity element bottom for \/.

Laws

x \/ bottom ≡ x

Corollary

x /\ bottom
  ≡⟨ identity ⟩
(x /\ bottom) \/ bottom
  ≡⟨ absorption ⟩
bottom

Methods

bottom :: a

class Lattice a => BoundedMeetSemiLattice a where

A meet-semilattice with an identity element top for /\.

Laws

x /\ top ≡ x

Corollary

x \/ top
  ≡⟨ identity ⟩
(x \/ top) /\ top
  ≡⟨ absorption ⟩
top

Methods

top :: a

joins :: (BoundedJoinSemiLattice a, Foldable f) => f a -> a

The join of a list of join-semilattice elements

meets :: (BoundedMeetSemiLattice a, Foldable f) => f a -> a

The meet of a list of meet-semilattice elements

fromBool :: BoundedLattice a => Bool -> a

True to top and False to bottom

type BoundedLattice a = (BoundedMeetSemiLattice a, BoundedJoinSemiLattice a)

Monoid wrappers

newtype Meet a

Monoid wrapper for meet-Lattice

Constructors

Meet

Fields

getMeet :: a

newtype Join a

Monoid wrapper for join-Lattice

Constructors

Join

Fields

getJoin :: a

Fixed points of chains in lattices

lfp :: (Eq a, BoundedJoinSemiLattice a) => (a -> a) -> a

Implementation of Kleene fixed-point theorem http://en.wikipedia.org/wiki/Kleene_fixed-point_theorem. Forces the function to be monotone.

lfpFrom :: (Eq a, BoundedJoinSemiLattice a) => a -> (a -> a) -> a

Implementation of Kleene fixed-point theorem http://en.wikipedia.org/wiki/Kleene_fixed-point_theorem. Forces the function to be monotone.

unsafeLfp :: (Eq a, BoundedJoinSemiLattice a) => (a -> a) -> a

Implementation of Kleene fixed-point theorem http://en.wikipedia.org/wiki/Kleene_fixed-point_theorem. Assumes that the function is monotone and does not check if that is correct.

gfp :: (Eq a, BoundedMeetSemiLattice a) => (a -> a) -> a

Implementation of Kleene fixed-point theorem http://en.wikipedia.org/wiki/Kleene_fixed-point_theorem. Forces the function to be antinone.

gfpFrom :: (Eq a, BoundedMeetSemiLattice a) => a -> (a -> a) -> a

Implementation of Kleene fixed-point theorem http://en.wikipedia.org/wiki/Kleene_fixed-point_theorem. Forces the function to be antinone.

unsafeGfp :: (Eq a, BoundedMeetSemiLattice a) => (a -> a) -> a

Implementation of Kleene fixed-point theorem http://en.wikipedia.org/wiki/Kleene_fixed-point_theorem. Assumes that the function is antinone and does not check if that is correct.