Algebra.Lattice.Lifted

Description

Documentation

data Lifted a

Graft a distinct bottom onto an otherwise unbounded lattice. As a bonus, the bottom will be an absorbing element for the meet.

Constructors

Bottom

Lift a

retractLifted :: BoundedJoinSemiLattice a => Lifted a -> a

Interpret Lifted a using the BoundedJoinSemiLattice of a.

foldLifted :: b -> (a -> b) -> Lifted a -> b

Similar to maybe, but for Lifted type.